requirements.txt
Appium-Python-Client==1.3.0
mysql-connector-python==8.0.31
pytest==7.1.3
selenium==3.141.0
requests==2.25.1